Root admin tenant id resolver added (for operations that affect tables with Multitenancy support).

+++ identity-server/src/main/java/com/hevelian/identity/server/tenants/resolvers/SessionTenantResolver.java
@@ -5,6 +5,6 @@ import com.hevelian.identity.server.tenants.TenantUtil;
 public class SessionTenantResolver implements CurrentTenantResolver<Long> {
     @Override
     public Long getCurrentTenantId() {
-        return (Long) TenantUtil.getCurrentTenantId();
+        return TenantUtil.getCurrentTenantId();
     }
 }
